Given authorization

+++ src/tasks/tasks.service.ts
@@ -1,4 +1,12 @@
-import { Injectable, GoneException, NotFoundException, Get, Param, ParseIntPipe } from '@nestjs/common';
+import { User } from './../auth/user.entity';
+import {
+  Injectable,
+  GoneException,
+  NotFoundException,
+  Get,
+  Param,
+  ParseIntPipe,
+} from '@nestjs/common';
 import { GetTasksFiltersDto } from 'src/DTO/get-task-filter.dto';
 import { InjectRepository, handleRetry } from '@nestjs/typeorm';
 import { TaskRepository } from './task.repository';
@@ -6,52 +14,44 @@ import { Task } from './task.entity';
 import { TaskStatus } from './task-status.enum';
 import { CreateTaskDto } from 'src/DTO/create-task.DTO';
 
-
 @Injectable()
 export class TasksService {
- 
-    constructor(
-        @InjectRepository(TaskRepository)
-        private taskRepository:TaskRepository,
-    ){}
-    async getTask(filterDto:GetTasksFiltersDto):Promise<Task[]>{
-        return this.taskRepository.getTasks(filterDto);
+  constructor(
+    @InjectRepository(TaskRepository)
+    private taskRepository: TaskRepository,
+  ) {}
+  async getTasks(filterDto: GetTasksFiltersDto, user: User): Promise<Task[]> {
+    return this.taskRepository.getTasks(filterDto, user);
+  }
+  async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
+    return this.taskRepository.createTask(createTaskDto, user);
+  }
+  async getTaskById(id: number, user: User): Promise<Task> {
+    const found = this.taskRepository.findOne({
+      where: { id, userId: user.id },
+    });
+    if (!found) {
+      throw new NotFoundException('task not found');
     }
-    async createTask(createTaskDto:CreateTaskDto):Promise<Task>{
-    
-        const {title,descricao}=createTaskDto;
-        const task = new Task();
-        task.title=title;
-        
-        task.descricao=descricao;
-        task.status=TaskStatus.IN_PROGRESS;
-        await task.save();
-        
-        return task;
-        }
-     async getTaskById(id:number):Promise<Task>{
-         const found = this.taskRepository.findOne(id);
-         if(!found){
-             throw new NotFoundException("task not found");
-         }
-         return found;
-     }
+    return found;
+  }
 
+  async deleteTask(id: number, user: User): Promise<void> {
+    const result = await this.taskRepository.delete({ id, userId: user.id });
+    console.log(result);
 
-   async deleteTask(id:number):Promise<void>{
-       const result= await this.taskRepository.delete(id);
-        console.log(result)
-
-        if(result.affected==0){
-            throw new NotFoundException('task not found');
-        }
+    if (result.affected == 0) {
+      throw new NotFoundException('task not found');
     }
-    async updateTaskStatus(id:number,status:TaskStatus):Promise<Task>{
-    const task=await this.getTaskById(id);
-           task.status=status;
-            await task.save();
-            return task;
-        }
-
-
+  }
+  async updateTaskStatus(
+    id: number,
+    user: User,
+    status: TaskStatus,
+  ): Promise<Task> {
+    const task = await this.getTaskById(id, user);
+    task.status = status;
+    await task.save();
+    return task;
+  }
 }
